Add new Anki & GSM stats page
Added a stats page with a Kanji grid to show all Kanji you have seen reading which are not in Anki

Not sure what else to add here but for now this is good!

This is its own page because it's Anki specific and requires Anki Connect, and I might think of other Anki specific stats to add in the future

Also removed the text replacements page.